requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
yfinance>=0.2.0
plotly>=5.17.0
duckduckgo-search>=4.0.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
langchain-core>=0.1.0
llama-index>=0.9.0
llama-index-embeddings-ollama>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-readers-database>=0.1.0
curl-cffi>=0.5.0
urllib3>=2.0.0
python-dateutil>=2.8.0
quiffen>=0.1.0
ollama>=0.1.0
pgvector>=0.3.0